feat: implement algorithm to find longest path of maze (not fully working)

+++ it_learns_mazes/environment.py
@@ -49,7 +49,7 @@ class MazeEnv(gym.Env):
         self.clock.tick(self.metadata["render_fps"])
 
     def _get_obs(self):
-        return np.array(self.game.rays)
+        return np.array(self.game.rays, dtype=np.float32)
 
     def step(self, action):
         terminated = False
